+++ hyperdbg/libhyperdbg/header/debugger.h
@@ -119,9 +119,6 @@ typedef struct _DEBUGGER_SYNCRONIZATION_EVENTS_STATE
 //				    Functions                   //
 //////////////////////////////////////////////////
 
-VOID
-InterpreterRemoveComments(char * CommandText);
-
 BOOLEAN
 ShowErrorMessage(UINT32 Error);